// client/app/api/identity/model/RegisterModel.ts
/**
 * GuanPlus.Identity.WebApi
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import * as models from './models';

export interface RegisterModel {
    /**
     * 显示名称
     */
    displayName?: string;

    /**
     * 代账公司名称
     */
    companyName?: string;

    /**
     * 手机或者邮箱
     */
    phoneOrEmail?: string;

    /**
     * Password
     */
    password?: string;

    /**
     * 验证码
     */
    verifyCode?: string;

    /**
     * 注册来源
     */
    source?: RegisterModel.SourceEnum;

    /**
     * 活动代码
     */
    campaign?: string;

    /**
     * Keyword
     */
    keyword?: string;

    /**
     * 推荐人Id
     */
    refereeId?: string;

    /**
     * 分享活动代码
     */
    code?: string;

    /**
     * 应用码
     */
    appCode?: string;

}
export namespace RegisterModel {
    export enum SourceEnum {
        WebPage = <any> 'WebPage',
        MobileWeb = <any> 'MobileWeb',
        Android = <any> 'Android',
        IOS = <any> 'IOS',
        DingDing = <any> 'DingDing'
    }
}